python3: Use disk_ids as a list and not dict_values object
In python3 the `.values()` function returns a `dict_values` object and it is not
iterable and generally not usable as `dict` is.  To fix these issues, just make
a list out of this in the specific place which uses it in such way.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>
(cherry picked from commit 8aa2f45f4dbcf4a031870f23e854d6848a7abca9)
Signed-off-by: Martin Kletzander <mkletzan@redhat.com>