Fail gracefully on rpmUtils missing
That function is designed in a way that it can fail gracefuylly and it would be
nicer not to fail unless it is necessary.  Not having the package should just
print a message and end with an error, not a stacktrace.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>
(cherry picked from commit 02975f3be1dce8f26ac1c3c6f82eda56d38cc005)
Signed-off-by: Martin Kletzander <mkletzan@redhat.com>