Only delete hidden keys if they exist in the object
This is mostly for the `internal` attribute that is automatically skipped for
all objects, even if they are not part of it.  I know there are no such objects
now, but it can be used later.

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>
(cherry picked from commit b6510e7e0b30a98df5df5dd70e177b8c60818870)
Signed-off-by: Martin Kletzander <mkletzan@redhat.com>